+++ src/builder.rs
@@ -132,7 +132,7 @@ pub struct BuildConfig<S> {
 pub struct DeviceBuilder<C = NoConnect, S = NoStore> {
     pub(crate) channel_size: usize,
     pub(crate) volatile_retention: usize,
-    pub(crate) store_retention: NonZeroUsize,
+    pub(crate) stored_retention: NonZeroUsize,
     pub(crate) store: S,
     pub(crate) connection_config: C,
     pub(crate) interfaces: Interfaces,
@@ -158,7 +158,7 @@ impl DeviceBuilder<NoConnect, NoStore> {
         Self {
             channel_size: DEFAULT_CHANNEL_SIZE,
             volatile_retention: DEFAULT_VOLATILE_CAPACITY,
-            store_retention: DEFAULT_STORE_CAPACITY,
+            stored_retention: DEFAULT_STORE_CAPACITY,
             writable_dir: None,
             interfaces: Interfaces::new(),
             connection_config: NoConnect,
@@ -276,8 +276,7 @@ impl<C> DeviceBuilder<C, NoStore> {
 
         self = self.writable_dir(path)?;
 
-        let mut store = SqliteStore::connect(path).await?;
-        store.set_max_items(self.store_retention).await?;
+        let store = SqliteStore::connect(path).await?;
 
         Ok(self.store(store))
     }
@@ -293,28 +292,22 @@ impl<C> DeviceBuilder<C, NoStore> {
             interfaces: self.interfaces,
             connection_config: self.connection_config,
             store,
-            store_retention: self.store_retention,
+            stored_retention: self.stored_retention,
             channel_size: self.channel_size,
             volatile_retention: self.volatile_retention,
             writable_dir: self.writable_dir,
         }
     }
+}
 
+impl<S> DeviceBuilder<NoConnect, S>
+where
+    S: StoredRetention,
+{
     /// Set the maximum number of elements that will be kept in the store
-    ///
-    /// if the provided size is zero, the default value of [`crate::builder::DEFAULT_STORE_CAPACITY`] will be used.
     #[instrument(skip(self))]
-    pub fn max_retention_items(mut self, size: usize) -> Self {
-        let Some(size) = NonZeroUsize::new(size) else {
-            debug!(
-                "Maintaining currebt retention store size: {}",
-                self.store_retention
-            );
-
-            return self;
-        };
-
-        self.store_retention = size;
+    pub fn max_retention_items(mut self, size: NonZeroUsize) -> Self {
+        self.stored_retention = size;
 
         self
     }
@@ -340,7 +333,7 @@ where
             store: self.store,
             channel_size: self.channel_size,
             volatile_retention: self.volatile_retention,
-            store_retention: self.store_retention,
+            stored_retention: self.stored_retention,
             writable_dir: self.writable_dir,
         }
     }
@@ -388,6 +381,13 @@ where
             sender,
             store,
         } = self.connection_config.connect(config).await?;
+
+        // set max retention items in the store
+        if let Some(retention) = store.get_retention() {
+            retention
+                .set_max_retention_items(self.stored_retention)
+                .await?;
+        }
 
         let client =
             DeviceClient::new(sender.clone(), rx_client, store.clone(), Arc::clone(&state));